Rename library to spatial3d for publishing to PyPI.
