+++ spatial3d/coordinate_axes.py
