+++ spatial3d/dual.py
