+++ spatial3d/edge.py
