+++ spatial3d/euler/__init__.py
