+++ spatial3d/exceptions.py
