+++ spatial3d/facet.py
