+++ spatial3d/intersection.py
