+++ spatial3d/mesh.py
