+++ spatial3d/parameters.py
