+++ spatial3d/quaternion.py
