+++ spatial3d/ray.py
