+++ spatial3d/swizzler.py
