+++ test/test_transform.py
@@ -1,8 +1,8 @@
 import math
 import unittest
 
-from spatial import Quaternion, Transform, Vector3
-from spatial.euler import Axes, Order
+from spatial3d import Quaternion, Transform, Vector3
+from spatial3d.euler import Axes, Order
 
 
 class TestTransform(unittest.TestCase):
